refactor: split sliding panel buttons into multiple files
