+++ packages/x-tailwindcss/src/x-tailwind-plugin/components/sliding-panel/buttons/hover.ts
@@ -1,0 +1,38 @@
+/**
+ * Returns the button hover variants for the component `sliding panel`.
+ *
+ * @returns The {@link CssStyleOptions} for the variant.
+ */
+export function slidingPanelButtonsHover() {
+  return {
+    '&.sliding-panel-show-buttons-on-hover': {
+      '.sliding-panel-button': {
+        opacity: 0
+      },
+      '&:not(.sliding-panel-at-start):hover': {
+        '.sliding-panel-button-left': {
+          opacity: 1,
+          pointerEvents: 'all'
+        }
+      },
+      '&:not(.sliding-panel-at-end):hover': {
+        '.sliding-panel-button-right': {
+          opacity: 1,
+          pointerEvents: 'all'
+        }
+      }
+    },
+    '&:not(.sliding-panel-show-buttons-on-hover):not(.sliding-panel-at-start)': {
+      '.sliding-panel-button-left': {
+        opacity: 1,
+        pointerEvents: 'all'
+      }
+    },
+    '&:not(.sliding-panel-show-buttons-on-hover):not(.sliding-panel-at-end)': {
+      '.sliding-panel-button-right': {
+        opacity: 1,
+        pointerEvents: 'all'
+      }
+    }
+  };
+}
