+++ packages/x-tailwindcss/src/x-tailwind-plugin/components/sliding-panel/buttons/positions.ts
@@ -1,0 +1,27 @@
+/**
+ * Returns the button position variants for the component `sliding panel`.
+ *
+ * @returns The {@link CssStyleOptions} for the variant.
+ */
+export function slidingPanelButtonsPositions() {
+  return {
+    '&.buttons-center': {
+      '.sliding-panel-button-left': {
+        transform: 'translateX(-50%)'
+      },
+      '.sliding-panel-button-right': {
+        transform: 'translateX(50%)'
+      }
+    },
+    '&.buttons-outside': {
+      '.sliding-panel-button-left': {
+        pointerEvents: 'all',
+        transform: 'translateX(-100%)'
+      },
+      '.sliding-panel-button-right': {
+        pointerEvents: 'all',
+        transform: 'translateX(100%)'
+      }
+    }
+  };
+}
